requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
frozenlist==1.7.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
multidict==6.5.0
openai==0.28.1
propcache==0.3.2
python-telegram-bot==20.7
pytz==2025.2
requests==2.32.4
rfc3986==1.5.0
setuptools==80.9.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wheel==0.45.1
yarl==1.20.1
